+++ suomen-vaylat-app/src/components/menus/hierarchical-layerlist/ThemeLayerList.jsx
@@ -64,6 +64,9 @@ const StyledMasterGroupHeader = styled.div`
     cursor: pointer;
     background-color: ${props => props.theme.colors.secondaryColor2};
     border-radius: ${props => props.isOpen ? "4px 4px 0px 0px": "4px"};
+    @-moz-document url-prefix() {
+        position: initial;
+    };
 `;
 
 const StyledLeftContent = styled.div`
@@ -161,7 +164,12 @@ export const ThemeLayerList = ({
 
     const { store } = useContext(ReactReduxContext);
 
-    const { channel, selectedTheme,  lastSelectedTheme, selectedThemeIndex} = useAppSelector((state) => state.rpc);
+    const {
+        channel,
+        selectedTheme, 
+        lastSelectedTheme,
+        selectedThemeIndex
+    } = useAppSelector((state) => state.rpc);
 
     const handleSelectGroup = (index, theme) => {
         selectGroup(store, channel, index, theme, lastSelectedTheme, selectedThemeIndex);
@@ -231,7 +239,12 @@ export const ThemeGroup = ({
             >
                 <div>
                     {themeImages[theme.id] && <StyledLayerGroupImage src={themeImages[theme.id]} alt=''/>}
-                    {strings.themelayerlist[theme.id].description !== null &&
+                    {strings.themelayerlist.hasOwnProperty(theme.id) && strings.themelayerlist[theme.id].title !== null &&
+                    <>
+                        <StyledSubHeader>{strings.themelayerlist[theme.id].title}</StyledSubHeader>
+                    </>
+                    }
+                    {strings.themelayerlist.hasOwnProperty(theme.id) && strings.themelayerlist[theme.id].description !== null &&
                     <>
                         <StyledSubHeader>{strings.themelayerlist[theme.id].title}</StyledSubHeader>
                         <StyledSubText>{strings.themelayerlist[theme.id].description}</StyledSubText>