+++ ReleaseNotes.md
@@ -1,5 +1,20 @@
 # Release Notes
 
+## 0.3.0
+
+For a full lis of changes see: https://github.com/finnishtransportagency/suomen-vaylat/milestone/2?closed=1
+
+* React-modal changed to own modal component
+* Fixed an issue for theme sharing
+* Fixed an issue for layers component
+* Fixed an issue for mobile Firefox
+* Added more translations
+* Set layer subgroups alphabetically on the front end side
+* Improve address search
+* UI tuning
+* GFI modal minimaze
+
+
 ## 0.2.0
 
 New UI.